Add a "fulcrum point" to d3.event for zoom+pan.
The fulcrum is mouse point at the initiation of the action.
The idea is that listeners to d3.behavior.zoom() and others
might like to use that point as a context for interaction.
The zoom-pan demo has been modified to use this so that
axes can be independently adjusted by interacting in the
plot margins.

However, there's a problem: after adjusting a single axis,
any adjustment to the other axis will include all of the
movements made while adjusting the original axis.